deps/v8/test/inspector/debugger/async-stack-created-frame-expected.txt

Checks created frame for async call chain

Running test: testPromise
foo1 (test.js:10:2)
-- Promise.then --
promise (test.js:20:14)
(anonymous) (expr.js:0:0)


Running test: testPromiseThen
foo1 (test.js:10:2)
-- Promise.then --
promiseThen (test.js:28:14)
(anonymous) (expr.js:0:0)

foo2 (test.js:14:2)
-- Promise.then --
promiseThen (test.js:29:14)
(anonymous) (expr.js:0:0)


Running test: testPromiseThenThen
foo1 (test.js:10:2)
-- Promise.then --
promiseThenThen (test.js:37:14)
(anonymous) (expr.js:0:0)

foo1 (test.js:10:2)
-- Promise.then --
promiseThenThen (test.js:38:14)
(anonymous) (expr.js:0:0)

foo2 (test.js:14:2)
-- Promise.then --
promiseThenThen (test.js:37:25)
(anonymous) (expr.js:0:0)


Running test: testPromiseResolve
foo1 (test.js:10:2)
-- Promise.then --
promiseResolve (test.js:44:27)
(anonymous) (expr.js:0:0)


Running test: testPromiseReject
foo1 (test.js:10:2)
-- Promise.catch --
promiseReject (test.js:48:31)
(anonymous) (expr.js:0:0)


Running test: testPromiseAll
foo1 (test.js:10:2)
-- Promise.then --
promiseAll (test.js:52:44)
(anonymous) (expr.js:0:0)


Running test: testPromiseRace
foo1 (test.js:10:2)
-- Promise.then --
promiseRace (test.js:56:45)
(anonymous) (expr.js:0:0)


Running test: testThenableJob1
foo1 (test.js:10:2)
-- Promise.then --
thenableJob1 (test.js:60:72)
(anonymous) (expr.js:0:0)


Running test: testThenableJob2
foo1 (test.js:10:2)
-- Promise.then --
thenableJob2 (test.js:64:57)
(anonymous) (expr.js:0:0)


Running test: testSetTimeouts
foo1 (test.js:10:2)
(anonymous) (test.js:72:25)
-- setTimeout --
(anonymous) (test.js:72:6)
-- setTimeout --
(anonymous) (test.js:71:4)
-- setTimeout --
setTimeouts (test.js:70:2)
(anonymous) (expr.js:0:0)
